snapper: Properly detect the type of a snapshot
Previously we were just hacking in the type of number to guess if it's
pre and post. What we're missing there is:
a). single snapshots exist
b). single snapshots can be taken manually

This completely breaks the assumption that was made, and can then create
incorrect entries for corresponding snapshots.

Signed-off-by: Eric Naim <dnaim@cachyos.org>